Modify Layout.jsx to Layout.tsx, put correct type on context. Change the height of the div that contain all the element to fit all the screen
