+++ e-com-training/src/layout/components/Footer.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 const Footer = () => {
     return (
-        <div></div>
+        <div className="absolute inset-x-0 bottom-0 h-16">FOOTER</div>
     )
 }
